Useless library.
